* moio/plugins/uis/MainFrame.py,   moio/plugins/uis/AddressBookDialog.py:  - conferma cancellazione contatto * moio/plugins/uis/AddressBookDialog.py:  - aggiunto all'interfaccia il bottone per la funzione importa

+++ moio/plugins/uis/MainFrame.py
@@ -535,6 +535,11 @@ class MainFrame(QFrame):
         destString = unicode(self.destinationComboBox.currentText())
         destIndex = self.destinationComboBox.findText(
                                 self.destinationComboBox.currentText())
+        result = QMessageBox.question(self,
+            u"Attenzione!",
+            u"Cancellare veramente il contatto?\n", 'Si','No')
+        if result == 1:
+            return
         try:
             self.book.deleteContact(destString)
         except BookError, be:
